Split library in two: main + core utilities
