+++ lerna.json
@@ -1,0 +1,8 @@
+{
+  "$schema": "node_modules/lerna/schemas/lerna-schema.json",
+  "version": "0.0.0",
+  "packages": [
+    "packages/*"
+  ],
+  "npmClient": "yarn"
+}
