+++ packages/mambajs-core/package.json
@@ -1,0 +1,43 @@
+{
+  "name": "@emscripten-forge/mambajs-core",
+  "version": "0.12.1",
+  "description": "Mambajs core utilities",
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
+  "scripts": {
+    "clean": "rimraf lib && rimraf tsconfig.tsbuildinfo",
+    "build": "yarn run clean && yarn run build:esbuild && yarn run build:types",
+    "build:types": "tsc --emitDeclarationOnly",
+    "build:esbuild": "node esbuild.config.js"
+  },
+  "files": [
+    "lib/*"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "keywords": [
+    "conda packages"
+  ],
+  "homepage": "https://github.com/emscripten-forge/mambajs",
+  "bugs": {
+    "url": "https://github.com/emscripten-forge/mambajs/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/emscripten-forge/mambajs"
+  },
+  "author": "Emscripten-forge developers",
+  "license": "MIT",
+  "dependencies": {
+    "@emscripten-forge/untarjs": "^5.3.0",
+    "yaml": "^2.7.0"
+  },
+  "devDependencies": {
+    "@esbuild-plugins/node-globals-polyfill": "^0.2.3",
+    "@esbuild-plugins/node-modules-polyfill": "^0.2.2",
+    "esbuild": "^0.24.0",
+    "rimraf": "^3.0.2",
+    "typescript": "^5"
+  }
+}
