+++ packages/mambajs-core/src/helper.ts
@@ -34,14 +34,6 @@ export interface IEmpackEnvMeta {
   packages: IEmpackEnvMetaPkg[];
 }
 
-export interface ISolveOptions {
-  ymlOrSpecs?: string | string[];
-  installedPackages?: ISolvedPackages;
-  pipSpecs?: string[];
-  channels?: string[];
-  logger?: ILogger;
-}
-
 /**
  * Shared libraries. list of .so files
  */
