+++ packages/mambajs/README.md
@@ -1,0 +1,1 @@
+# Run js-rattler in a web-worker
