+++ tsconfig.eslint.json
@@ -1,7 +1,7 @@
 {
     "extends": "./tsconfig.json",
     "include": [
-      "src/**/*.ts"
+      "packages/**/*.ts"
     ],
     "exclude": [
       "node_modules",
