+++ tsconfig.json
@@ -6,29 +6,15 @@
     "esModuleInterop": true,
     "incremental": true,
     "jsx": "react",
-    "module": "es2020",
+    "module": "esnext",
     "moduleResolution": "node",
-    "noEmitOnError": false,
+    "noEmitOnError": true,
     "noImplicitAny": false,
     "noUnusedLocals": true,
     "preserveWatchOutput": true,
     "resolveJsonModule": true,
     "strict": true,
     "strictNullChecks": true,
-    "strictPropertyInitialization": false,
-    "downlevelIteration": true,
-    "target": "ES2020",
-    "types": ["node", "jest"],
-    "skipLibCheck": true,
-    "lib": ["ES2020", "WebWorker", "DOM"],
-    "allowJs": true,
-    "outDir": "lib",
-    "rootDir": "src",
-    "baseUrl": "./src",
-    "paths": {
-      "*": ["*"]
-    }
-  },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules"]
+    "target": "ES2019"
+  }
 }
